+++ scripts/build.sh
@@ -3,8 +3,7 @@ set -eu
 
 # https://node.green/#ES2022
 # https://kangax.github.io/compat-table/es2016plus
-TARGET='node16.9.0'
-VERSION=\""$npm_package_version"\"
+TARGET='node20.10.0'
 
 # build .d.ts
 tsc --build src/
@@ -12,7 +11,7 @@ tsc --build src/
 cp -f dist/index.d.ts dist/index.d.cts
 
 # build ESM
-esbuild 'src/**/*.ts' --target=$TARGET --define:VERSION=$VERSION --outdir=dist --log-level=warning
+esbuild 'src/**/*.ts' --target=$TARGET --outdir=dist --log-level=warning
 
 # build CommonJS (fallback)
 esbuild src/index.ts --bundle --target=$TARGET --platform=node > dist/index.cjs